requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.4
pytest==8.3.5
playwright==1.49.1
openai==1.82.0
Pillow==11.1.0
mkdocs==1.6.1
mkdocs-material==9.6.7
mkdocs-redirects==1.2.2
mkdocs-static-i18n==1.3.0
pymdown-extensions==10.14.3
